report common
