+++ src/offchain_service.rs
@@ -2,7 +2,8 @@ use std::{collections::HashMap, env, sync::Arc};
 
 use crate::{
     app_state::AppState, consts::GOOGLE_CHAT_REPORT_SPACE_URL,
-    events::push_notifications::subscribe_device_to_topic, AppError,
+    events::push_notifications::subscribe_device_to_topic,
+    posts::report_post::repost_post_common_impl, AppError,
 };
 use anyhow::{Context, Result};
 use axum::extract::State;
@@ -36,70 +37,6 @@ impl OffChain for OffChainService {
         let shared_state = self.shared_state.clone();
         let request = request.into_inner();
 
-        let text_str = format!(
-            "reporter_id: {} \n publisher_id: {} \n publisher_canister_id: {} \n post_id: {} \n video_id: {} \n reason: {} \n video_url: {}",
-            request.reporter_id, request.publisher_id, request.publisher_canister_id, request.post_id, request.video_id, request.reason, request.video_url
-        );
-
-        let data = json!({
-            "cardsV2": [
-            {
-                "cardId": "unique-card-id",
-                "card": {
-                    "sections": [
-                    {
-                        "header": "Report Post",
-                        "widgets": [
-                        {
-                            "textParagraph": {
-                                "text": text_str
-                            }
-                        },
-                        {
-                            "buttonList": {
-                                "buttons": [
-                                    {
-                                    "text": "View video",
-                                    "onClick": {
-                                        "openLink": {
-                                        "url": request.video_url
-                                        }
-                                    }
-                                    },
-                                    {
-                                    "text": "Ban Post",
-                                    "onClick": {
-                                        "action": {
-                                        "function": "goToView",
-                                        "parameters": [
-                                            {
-                                            "key": "viewType",
-                                            "value": format!("{} {}", request.publisher_canister_id, request.post_id),
-                                            }
-                                        ]
-                                        }
-                                    }
-                                    }
-                                ]
-                            }
-                        }
-                        ]
-                    }
-                    ]
-                }
-            }
-            ]
-        });
-
-        let res = send_message_gchat(GOOGLE_CHAT_REPORT_SPACE_URL, data).await;
-        if res.is_err() {
-            log::error!("Error sending data to Google Chat: {:?}", res);
-            return Err(tonic::Status::new(
-                tonic::Code::Unknown,
-                "Error sending data to Google Chat",
-            ));
-        }
-
         let user_principal = Principal::from_text(&request.reporter_id)
             .map_err(|_| tonic::Status::new(tonic::Code::Unknown, "Invalid reporter id"))?;
         let user_canister_id = shared_state
@@ -112,7 +49,12 @@ impl OffChain for OffChainService {
             Principal::from_text(&request.publisher_canister_id).map_err(|_| {
                 tonic::Status::new(tonic::Code::Unknown, "Invalid publisher canister id")
             })?;
-        let post_report_request = crate::posts::ReportPostRequest {
+        let publisher_principal = Principal::from_text(&request.publisher_id).map_err(|_| {
+            tonic::Status::new(tonic::Code::Unknown, "Invalid publisher canister id")
+        })?;
+        let post_report_request = crate::posts::report_post::ReportPostRequestV2 {
+            publisher_principal,
+            report_mode: crate::posts::report_post::ReportMode::Web,
             canister_id: publisher_canister_id,
             post_id: request
                 .post_id
@@ -124,15 +66,11 @@ impl OffChain for OffChainService {
             reason: request.reason,
         };
 
-        let qstash_client = shared_state.qstash_client.clone();
-        qstash_client
-            .publish_report_post(post_report_request)
+        repost_post_common_impl(shared_state.clone(), post_report_request)
             .await
-            .map_err(|_| {
-                tonic::Status::new(
-                    tonic::Code::Unknown,
-                    "Error sending data to qstash. request from leptos ssr",
-                )
+            .map_err(|e| {
+                log::error!("Failed to report post: {}", e);
+                tonic::Status::new(tonic::Code::Unknown, "Failed to report post")
             })?;
 
         Ok(tonic::Response::new(Empty {}))
